# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardSummaryMode = typeof DashboardSummaryMode[keyof typeof DashboardSummaryMode];


export const DashboardSummaryMode = {
  demo: 'demo',
  live: 'live',
} as const;

export type DashboardSummaryLiveStatus = typeof DashboardSummaryLiveStatus[keyof typeof DashboardSummaryLiveStatus];


export const DashboardSummaryLiveStatus = {
  connected: 'connected',
  demo: 'demo',
} as const;

export interface DashboardSummary {
  mode: DashboardSummaryMode;
  liveStatus: DashboardSummaryLiveStatus;
  balance: number;
  dayPnl: number;
  dayPnlPercent: number;
  riskScore: number;
  riskLabel: string;
  activePositions: number;
  protectedCapital: number;
  policyStatus: string;
  lastSyncedAt: string;
  topSymbol: string;
  topSymbolChange: number;
}

export type MarketSnapshotTrend = typeof MarketSnapshotTrend[keyof typeof MarketSnapshotTrend];


export const MarketSnapshotTrend = {
  bullish: 'bullish',
  neutral: 'neutral',
  bearish: 'bearish',
} as const;

export interface MarketSnapshot {
  symbol: string;
  price: number;
  change24h: number;
  volatility: number;
  trend: MarketSnapshotTrend;
  signal: string;
  confidence: number;
  support: number;
  resistance: number;
  volume24h: number;
  candles: number[];
}

export type PortfolioAssetTone = typeof PortfolioAssetTone[keyof typeof PortfolioAssetTone];


export const PortfolioAssetTone = {
  gold: 'gold',
  blue: 'blue',
  green: 'green',
  purple: 'purple',
} as const;

export interface PortfolioAsset {
  asset: string;
  value: number;
  allocation: number;
  change: number;
  tone: PortfolioAssetTone;
}

export interface PortfolioSummary {
  totalValue: number;
  availableCash: number;
  unrealizedPnl: number;
  leverage: number;
  concentration: number;
  assets: PortfolioAsset[];
}

export type ActivityEntryKind = typeof ActivityEntryKind[keyof typeof ActivityEntryKind];


export const ActivityEntryKind = {
  analysis: 'analysis',
  policy: 'policy',
  execution: 'execution',
  sync: 'sync',
  blocked: 'blocked',
} as const;

export type ActivityEntryStatus = typeof ActivityEntryStatus[keyof typeof ActivityEntryStatus];


export const ActivityEntryStatus = {
  completed: 'completed',
  awaiting: 'awaiting',
  blocked: 'blocked',
} as const;

export interface ActivityEntry {
  id: number;
  kind: ActivityEntryKind;
  title: string;
  detail: string;
  time: string;
  status: ActivityEntryStatus;
  /** @nullable */
  symbol: string | null;
}

export interface ExecutionPolicy {
  maxTradeSize: number;
  maxLeverage: number;
  dailyLossLimit: number;
  requireConfirmation: boolean;
  allowedSymbols: string[];
  withdrawalsBlocked: boolean;
}

export interface ExecutionPolicyInput {
  /** @minimum 1 */
  maxTradeSize: number;
  /** @minimum 1 */
  maxLeverage: number;
  /** @minimum 1 */
  dailyLossLimit: number;
  requireConfirmation: boolean;
  allowedSymbols: string[];
}

export type AgentAnalyzeInputIntent = typeof AgentAnalyzeInputIntent[keyof typeof AgentAnalyzeInputIntent];


export const AgentAnalyzeInputIntent = {
  observe: 'observe',
  simulate: 'simulate',
  prepare: 'prepare',
} as const;

export interface AgentAnalyzeInput {
  /** @minLength 1 */
  prompt: string;
  symbol: string;
  intent: AgentAnalyzeInputIntent;
}

export type AgentRunAction = typeof AgentRunAction[keyof typeof AgentRunAction];


export const AgentRunAction = {
  buy: 'buy',
  sell: 'sell',
  hold: 'hold',
} as const;

export type AgentRunSide = typeof AgentRunSide[keyof typeof AgentRunSide];


export const AgentRunSide = {
  LONG: 'LONG',
  SHORT: 'SHORT',
  FLAT: 'FLAT',
} as const;

export type AgentRunStatus = typeof AgentRunStatus[keyof typeof AgentRunStatus];


export const AgentRunStatus = {
  ready: 'ready',
  blocked: 'blocked',
  executed: 'executed',
} as const;

export interface AgentRun {
  id: number;
  symbol: string;
  headline: string;
  thesis: string;
  action: AgentRunAction;
  side: AgentRunSide;
  notional: number;
  entry: number;
  stop: number;
  target: number;
  confidence: number;
  riskScore: number;
  checks: string[];
  status: AgentRunStatus;
  createdAt: string;
}

export interface ExecuteActionInput {
  runId: number;
  confirmation: boolean;
}

export interface ExecutionResult {
  success: boolean;
  message: string;
  orderId: string;
  filledPrice: number;
  notional: number;
  riskAfter: number;
  activity: ActivityEntry;
}

export type GetMarketSnapshotParams = {
symbol?: string;
};

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};