# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntryKind } from './activityEntryKind';
import type { ActivityEntryStatus } from './activityEntryStatus';

export interface ActivityEntry {
  id: number;
  kind: ActivityEntryKind;
  title: string;
  detail: string;
  time: string;
  status: ActivityEntryStatus;
  /** @nullable */
  symbol: string | null;
}