# lib/api-zod/src/generated/types/agentAnalyzeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentAnalyzeInputIntent } from './agentAnalyzeInputIntent';

export interface AgentAnalyzeInput {
  /** @minLength 1 */
  prompt: string;
  symbol: string;
  intent: AgentAnalyzeInputIntent;
}