# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunAction } from './agentRunAction';
import type { AgentRunSide } from './agentRunSide';
import type { AgentRunStatus } from './agentRunStatus';

export interface AgentRun {
  id: number;
  symbol: string;
  headline: string;
  thesis: string;
  action: AgentRunAction;
  side: AgentRunSide;
  notional: number;
  entry: number;
  stop: number;
  target: number;
  confidence: number;
  riskScore: number;
  checks: string[];
  status: AgentRunStatus;
  createdAt: string;
}