# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryLiveStatus } from './dashboardSummaryLiveStatus';
import type { DashboardSummaryMode } from './dashboardSummaryMode';

export interface DashboardSummary {
  mode: DashboardSummaryMode;
  liveStatus: DashboardSummaryLiveStatus;
  balance: number;
  dayPnl: number;
  dayPnlPercent: number;
  riskScore: number;
  riskLabel: string;
  activePositions: number;
  protectedCapital: number;
  policyStatus: string;
  lastSyncedAt: string;
  topSymbol: string;
  topSymbolChange: number;
}