# lib/api-zod/src/generated/types/executeActionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteActionInput {
  runId: number;
  confirmation: boolean;
}