# lib/api-zod/src/generated/types/executionPolicy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecutionPolicy {
  maxTradeSize: number;
  maxLeverage: number;
  dailyLossLimit: number;
  requireConfirmation: boolean;
  allowedSymbols: string[];
  withdrawalsBlocked: boolean;
}