# lib/api-zod/src/generated/types/executionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEntry } from './activityEntry';

export interface ExecutionResult {
  success: boolean;
  message: string;
  orderId: string;
  filledPrice: number;
  notional: number;
  riskAfter: number;
  activity: ActivityEntry;
}