# lib/api-zod/src/generated/types/marketSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketSnapshotTrend } from './marketSnapshotTrend';

export interface MarketSnapshot {
  symbol: string;
  price: number;
  change24h: number;
  volatility: number;
  trend: MarketSnapshotTrend;
  signal: string;
  confidence: number;
  support: number;
  resistance: number;
  volume24h: number;
  candles: number[];
}