# lib/api-zod/src/generated/types/portfolioSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Orbit Binance Agent OS API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioAsset } from './portfolioAsset';

export interface PortfolioSummary {
  totalValue: number;
  availableCash: number;
  unrealizedPnl: number;
  leverage: number;
  concentration: number;
  assets: PortfolioAsset[];
}